Perform fine-grained SPIR-V capability checks (#129)
* Make variable pointers an optional extension

Enable whichever variable pointer features are supported by the Vulkan
device. This allows Clvk to run on top of SwiftShader for kernels that
do not require variable pointers.

* Check SPIR-V capabilities against device

This produces more useful errors (currently just to the logging
system) when a SPIR-V module produced by Clspv uses capabilities that
are not supported by the Vulkan device.