Remove deleted move special member functions
Fixes a build error occurring with vectors of refcounted_holder
objects. The push_back/emplace_back methods require the element type
to be MoveInsertable, and will try to use the move constructor if it
is declared. This produces an error if the move constructor is
deleted, but if it is left undeclared the compiler can use the copy
constructor instead.