Fix use of vkGetPhysicalDeviceFeatures2 for Vulkan 1.0
The Vulkan loader may report a newer API version than the underlying
Vulkan device, so we cannot rely on the version reported by
`vkGetInstanceVersion` to determine whether or not to enable the
extension. Instead, enable the extension if it is present, check the
device's API version at the point we use the function, and get the KHR
suffixed variant via `vkGetInstanceProcAddr` if we are on Vulkan 1.0.