Add memory barriers for async program builds
Make the program device status atomic, so that when build completes
the background thread can synchronize with a host thread that is
checking the status.

Set the device status *after* calling the completion callback.

The combination of these two changes fixes a TSAN issue for the
async_build test in CTS.