Fix clFinish when called just after clFlush
Always keep the last event produced as part of sending work to the
executor in the queue. Wait on that event in clFinish.

The test is not bullet-proof (hard to guarantee that the thread
about to call clFinish is blocked without blocking) but in my tests
failed 100% of the time before the fix and passed 100% of the time
after.

Fixes #241

Signed-off-by: Kévin Petit <kpet@free.fr>