Support creating an image with a host pointer
This adds a deferred image initialization step that is executed
immediately prior to the first device command that uses the image. The
contents of the host pointer are copied to a staging buffer during
image creation, and then copied to the image on the device queue that
is receiving the first command.

This initialization step also performs an image layout transitions,
which fixes #108.