Add support for pipeline cache serialization
If the user sets the `CLVK_CACHE_DIR` environment variable to a
non-empty value, clvk will write Vulkan pipeline data to a per-device
file in that directory during teardown. On the next run, clvk will
load pipeline cache data from this directory and use it to seed the
initial pipeline cache.

This also changes the Vulkan pipeline cache to be device-scope, rather
than program scope.