Use per-module Vulkan pipeline caches (#282)
* Use per-module Vulkan pipeline caches

Compute the SHA-1 hash of the SPIR-V module and use that as a key for
both in-memory and serialized caches of Vulkan pipelines caches.

Perform these cache lookups during program build. If there was a cache
hit, we know that the SPIR-V has previously been validated, and so we
can skip validation this time round. This significantly improves
clBuildProgram() performance when creating programs from binaries.

* Add tests for SHA-1 implementation

* Add SHA-1 tests to CI coverage

* Use void* for cvk_sha1 data parameter