Enable support for long vectors
No regressions on patched CTS (with long vectors disabled in most
tests); some more unpatched tests passing \o/.

Support is far from complete in clspv but good enough for some
applications. Enabling it so that those work out of the box. The
downside is that clspv will now crash or produce not-for-user
error messages when hitting unsupported features.

Contributes to #52

Signed-off-by: Kévin Petit <kpet@free.fr>