fix for local buffer argument in kernel prototype (#336)
* fix for local buffer argument in kernel prototype

In OpenCL CTS 'vstore_local', we have this pattern:

for (args - last_arg) {
  clSetKernelArg(arg)
}
for (various alignmentOffset){
  clSetKernelArg(last_arg, alignmentOffset);
  clEnqueueNDRangeKernel();
}

When we clSetKernelArg the second time. As the kernel has been enqueue
once, we will create a new cvk_kernel_argument_values in
cvk_kernel::set_arg. And we will lose information about the local
buffer argument, which led to a local buffer with the wrong size.

Let's copy the needed information in order to ensure that the local
buffer will always have the proper size.

* add test