Split region that are bigger than vklimits (#359)
* Split region that are bigger than vklimits

When a region is bigger than the vulkan limits, we need to split it in
smaller region to execute it.
Do that only if we are using 'region_offset' and 'region_group_offset'
otherwise it won't work properly.
Add a test for it. The test is using a export function that is not
part of OpenCL and thus not really documented that allow to change the
vulkan limits. It enables the test to make sense without having to
know the physical limits and to run in a short amount of time. This
function is an extern C function not to have to consider the mandling
in 'exports.map'. Shall not be used appart from this test.

* fix formating

* reset vklimits at the end of the test

* fix wrong is_valid_device usage

* various fixes following kpet feedback

* switch from macro loops to rec fct

* add CLVK_UNIT_TESTING on some jobs

* fix formating