Fix 65fef82 (split of region bigger than vklimits)
Looking for region_offset and region_group_offset to know if we can
split the region is not enough as a test using only one of them can
still be split. But it will not be accepted by clvk with the current
test.
Instead let's looking into the kernel options to choose whether we try
it or not.

Also round_up the size of the workgroup to the local size in order to
keep region uniform.