clean api_tests output
api_tests output was containing the null character ('\0').
When you store the output of the test in a file, this character makes
some system (like the browser chrome) consider the file as binary
rather than text.
trim this null character in order to let chrome open the file as text.