tests: api: images.cpp: remove include of utils.hpp
utils.hpp has been added to use ARRAY_SIZE at some point in #387.
In the end, ARRAY_SIZE was not used but utils.hpp was still there.

Remove it as it is not needed, and on top of that, it is making the
clspv ci fail, as we are just lucky in clvk that the build system
managed to find "vulkan/vulkan.h" (which is a dependency of
"utils.hpp") while it should not be needed to build cl tests.