allow smaller batch and early flush
add more config values to configure clvk:
- max_first_batch_size
- max_group_size
- max_first_group_size

Those configs have default value on a device level

Keep track of the number of batch enqueued to be able to have a
smaller one when no batch has been enqueued to allow the GPU to start
computing earlier.
Keep track of the number of group sent to the executor to be have to
have early flush of smaller group to allow an early execution of
commands.