refactor profiling into event
It will help to introduce timeline semaphore
When implementing dependency management through timeline semaphores,
it introduces a change on which thread is submitting commands, and
which thread is actually updated the event as being complete (through
a wait on a event).

Also profiling timestamps are really about the time when an event
changed status. It makes more sense to me, to have the profiling logic
inside the event itself, when we are changing its status.