Handle feature queries for promoted Vulkan extensions
#435 correctly started requiring shaderSubgroupExtendedTypes
to report cl_khr_subgroups as supported.

The extension that introduced it was, correctly, only being requested
for versions of Vulkan prior to 1.2 where it was promoted to a core
feature.

The feature query code only performed queries for features provided by
an extension which meant that, on a 1.2+ implementation, the feature was
reported as VK_FALSE and thus cl_khr_subgroups was no longer reported.

Feature queries are now run as well for features from extensions that
have been promoted to a core Vulkan version.

Signed-off-by: Kévin Petit <kpet@free.fr>