fix test in cts when clReleaseCommandQueue is called before clFinish
When a queue is release before the execution of all commands, the
queue will be destroyed in `cvk_executor_thread::executor` when
releasing the holder on the queue. But at that point, the executor
will have the lock. When calling wait_idle here, we will never be
able to take the lock, generating the deadlock.