refactor profiling into event (#424)
* refactor profiling into event

It will help to introduce timeline semaphore
When implementing dependency management through timeline semaphores,
it introduces a change on which thread is submitting commands, and
which thread is actually updated the event as being complete (through
a wait on a event).

Also profiling timestamps are really about the time when an event
changed status. It makes more sense to me, to have the profiling logic
inside the event itself, when we are changing its status.

* kpet feedback

* fix formatting and kpet feedback

* command type + set_profiling_info return CL_SUCCESS

* remove m_finish_event and add cvk_executor_thread::wait_idle

* remove unneeded diff

* fix formatting

* fix test in cts when clReleaseCommandQueue is called before clFinish

When a queue is release before the execution of all commands, the
queue will be destroyed in `cvk_executor_thread::executor` when
releasing the holder on the queue. But at that point, the executor
will have the lock. When calling wait_idle here, we will never be
able to take the lock, generating the deadlock.

* remove unneeded `m_running=true`

* trace wait_idle

* Update src/queue.hpp

Co-authored-by: Kévin Petit <kpet@free.fr>