rework api_tests image write invalidate
The test is known to be a little flaky.
This PR is changing a bit the nature of the test.
It is checking that the map region is not written by clvk before being
given to the user. But it does not check for other copies that could
happen inside clvk.