add option to disable dependency on LLVM-SPIRV-Translator
Keep it enabled by default
According to the OpenCL specification, just removing it from the list
returned by CL_DEVICE_IL_VERSION should be enough.

This is useful for ChromeOS to remove dependency on
LLVM-SPIRV-Translator as it is not strictly needed for OpenCL
compliance and we do not have a use for it.