remove unneeded build in satisfy_data_dependencies
building the command in satisfy_data_dependencies is not needed. The
command is enqueue just after. During the enqueue the command will be
either built as a batchable command (in
cvk_command_queue::enqueue_command), or built when it will be added to
a batch (in cvk_command_batch::add_command).