fix free of descriptor sets in cvk_kernel_argument_values
We used to free descriptor sets when cvk_kernel_argument_values is
destroyed. But as cvk_kernel has a ref on it, it is only freed when
clReleaseKernel is called.

To be able to free them earlier, keep a refcount between
retain_resources and release_resources. When the counter reach zero,
free the descriptor sets, and set m_is_enqueued to false to make sure
we reallocate the descriptor set if we need to use them once again.

As descriptor sets can be allocated after being freed, reset
them to VK_NULL_HANDLE.