fix print when forgetting vector modifier
According to the OpenCL specification, when printing a vector if the
modifier is not used, it is an undefined behavior.
The current implement prints the first element only:
"%v4u", (1, 2, 3, 4) => '1, 1, 1, 1'

But in our case we can figure out the element size using information
coming from clspv. Use it only when we are in this undefined behavior
to print something making more sense:
"%v4u", (1, 2, 3, 4) => '1, 2, 3, 4'