fix subgroup support & add partial support of cl_intel_subgroups
This PR fix an issue on Intel device where the subgroup size used by
the compiler is not the default one exposed by the vulkan driver
through `m_subgroup_properties.subgroupSize`. It leads to error when
running CTS `test_api sub_group_dispatch` as the expected number of
subgroup does not match the one seen during execution.

To fix that, this PR uses
`VkPipelineShaderStageRequiredSubgroupSizeCreateInfo` to force the
subgroup size used by the compiler.

Doing so, it breaks TFLite because it assumes that the subgroup size
is always 16. Which used to work by luck because it was the size used
by the compiler for the kernel using subgroups on application I have
tried.
But now the default size exposed by vulkan is 32, thus kernels are
compiled with subgroup size of 32.

To avoid breaking TFLite, this PR implements a subpart of
`cl_intel_subgroups` which is enough for TFLite to use this extension
and specify the subgroup size inside the kernel using
`__attribute__((intel_reqd_sub_group_size))`.

This PR also adds 2 config variable. One to set a prefered subgroup
size to use if nothing is specified. Another to set a force subgroup
size overriding everything. This aims at allowing easy debug if/when
we will be facing other issues with subgroup size.