implement dynamic managment of batch parameters (#621)
Add an option to manage dynamicaly batch parameters. The aim is to get
the best parameter for an application. Because many things can happen
in the system, we are always updating to take any change of load in
account.
This algorithm has been tested on chromeos with various workloads and
produce great improvement performance wise.

This new option is disabled by default.