fix lifetime of cvk_entry_point (#721)
cvk_entry_point are passed to other structures as a pointer. But it
can be deleted at any point when do_build it called.

To avoid invalid access, make entry point a shared_ptr in the program
map.

This has been detected running the compiler suite of the CTS on
ChromeOS (flaky behaviour).