Add config option to reuse descriptor_set
On Mali devices, using the ARM DDK, some application quickly get to a
point where the driver consider the memory as fragemented and it not
capable of allocated new descriptor_set anymore.

To avoid that, do not free descriptor_set, reuse ones that have been
release from the application point of view.