fix do_build_inner when compiling a program twice
test_compiler get_linked_program_info_kernel_names is failing because
the test tries to compile the same program twice with different build
option (changing a define).

The second time, clvk tries to compile it with `-x ir` because
`m_binary_type == CL_PROGRAM_BINARY_TYPE_COMPILED_OBJECT`, which leads
to trying to compile a CL source interpret as a IR source.

Fix do_build_inner by making source we are not in the case of trying
to compile a CL source.