remove cvk_command_combine in favor of cvk_event_combine (#526)
cvk_command_combine prevents optimization where commands could be batch.
It is needed to have a single event representing the execution of the
group of commands. This PR creates a new event (cvk_event_combine)
which allows to have a single event grouping several events, thus
keeping the single event for the group of commands.

It also allows to make do_action a private function, thus having it
called only in one place. This is very convenient to implement future
optimization (using VkSemaphore).